Rename endGameConditions into endGame
It makes the import statement in App.js look nicer.
